backend/requirements_versioned.txt
aiofiles==24.1.0
aiosmtplib==2.0.2
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anthropic==0.40.0
anyio==4.3.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==2.4.1
asyncpg==0.29.0
attrs==24.2.0
autoevals==0.0.109
bcrypt==4.1.2
blinker==1.9.0
boto3==1.35.70
botocore==1.35.70
braintrust==0.0.176
braintrust_core==0.0.54
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
clickhouse-connect==0.8.7
cryptography==42.0.5
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.110.0
fastapi-mail==1.4.0
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-oauth==0.16.0
idna==3.6
iniconfig==2.0.0
ipython==8.22.2
isodate==0.7.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langsmith==0.2.2
Levenshtein==0.26.1
lxml==5.3.0
lz4==4.3.3
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
more-itertools==10.5.0
numpy==1.26.4
openai==1.30.2
openpyxl==3.1.5
orjson==3.10.12
packaging==24.0
pandas==2.2.2
parso==0.8.3
partialjson==0.0.8
pdfminer.six==20240706
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pwdlib==0.2.0
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
sentry-sdk==2.17.0
simple-salesforce==1.12.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.12.3
snowflake-sqlalchemy==1.7.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
sqlparse==0.5.1
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.36.3
stripe==11.4.1
text-unidecode==1.3
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.66.4
traitlets==5.14.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
XlsxWriter==3.2.0
zeep==4.3.1
zstandard==0.23.0
